Rename GCMTUpToLatest to GCMTUpdate and delete original GCMTUpdate
